// src/models/User.ts
export interface User {
    id?: string;
    createdAt?: string;
    email?: string;
    password?: string;
    status?: number;
    name?: string;
    lastname?: string;
    isDeleted?: number;
    profileImageUrl?: string;
    birthdate?: string;
    role?: string;
    phoneNumber?: string;
    updatedAt?: string;
}
  